pyproject.toml
[project]
name = "fastmcp-client"
version = "0.1.0"
description = "FastMCP client for interacting with MCP servers through Claude AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "fastmcp>=2.9.2",
    "python-dotenv>=1.1.1",
]

[tool.setuptools]
py-modules = ["client", "multi_client"]
